Refactor game logic into modular hooks
